fix filesize > 2BG ref #2572 https://forge.in2p3.fr/issues/2572

+++ adaptors/jsaga-adaptor-srm/src/fr/in2p3/jsaga/adaptor/data/SRM22FileAttributes.java
@@ -49,7 +49,7 @@ public class SRM22FileAttributes extends FileAttributes {
 
     public long getSize() {
         if (m_entry.getSize() != null) {
-            return m_entry.getSize().intValue();
+            return m_entry.getSize().longValue();
         } else {
             return SIZE_UNKNOWN;
         }
